+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_scheduler.py
@@ -14,14 +14,12 @@ self._to_observe is protected by an asyncio Lock
 """
 
 import asyncio
-import contextlib
 import functools
 import logging
-from asyncio import Lock, Queue, Task, sleep
-from dataclasses import dataclass, field
-from typing import Optional, Union
+from asyncio import sleep
+from dataclasses import dataclass
+from typing import Optional
 
-from fastapi import FastAPI
 from models_library.basic_types import PortInt
 from models_library.projects import ProjectID
 from models_library.projects_networks import DockerNetworkAlias
@@ -41,7 +39,6 @@ from .....models.schemas.dynamic_services import (
     DynamicSidecarStatus,
     RunningDynamicServiceDetails,
     SchedulerData,
-    ServiceName,
 )
 from ...api_client import DynamicSidecarClient, get_dynamic_sidecar_client
 from ...docker_api import (
@@ -56,7 +53,7 @@ from ...errors import (
     DynamicSidecarError,
     DynamicSidecarNotFoundError,
 )
-from .._abc import SchedulerInternalsInterface, SchedulerPublicInterface
+from .._abc import SchedulerPublicInterface
 from ._events_utils import (
     service_push_outputs,
     service_remove_containers,
@@ -64,6 +61,7 @@ from ._events_utils import (
     service_save_state,
 )
 from ._observer import observing_single_service
+from ._scheduler_mixin import SchedulerInternalsMixin
 
 logger = logging.getLogger(__name__)
 
@@ -72,25 +70,7 @@ _DISABLED_MARK = object()
 
 
 @dataclass
-class Scheduler(  # pylint: disable=too-many-instance-attributes
-    SchedulerPublicInterface, SchedulerInternalsInterface
-):
-
-    app: FastAPI
-
-    _lock: Lock = field(default_factory=Lock)
-    _to_observe: dict[ServiceName, SchedulerData] = field(default_factory=dict)
-    _service_observation_task: dict[
-        ServiceName, Optional[Union[asyncio.Task, object]]
-    ] = field(default_factory=dict)
-    _keep_running: bool = False
-    _inverse_search_mapping: dict[NodeID, str] = field(default_factory=dict)
-    _scheduler_task: Optional[Task] = None
-    _cleanup_volume_removal_services_task: Optional[Task] = None
-    _trigger_observation_queue_task: Optional[Task] = None
-    _trigger_observation_queue: Queue = field(default_factory=Queue)
-    _observation_counter: int = 0
-
+class Scheduler(SchedulerInternalsMixin, SchedulerPublicInterface):
     def toggle_observation(self, node_uuid: NodeID, disable: bool) -> bool:
         """
         returns True if it managed to enable/disable observation of the service
@@ -566,69 +546,3 @@ class Scheduler(  # pylint: disable=too-many-instance-attributes
                 logger.exception(
                     "Unexpected error while cleaning up pending volume removal services"
                 )
-
-    async def start(self) -> None:
-        # run as a background task
-        logger.info("Starting dynamic-sidecar scheduler")
-        self._keep_running = True
-        self._scheduler_task = asyncio.create_task(
-            self._run_scheduler_task(), name="dynamic-scheduler"
-        )
-        self._trigger_observation_queue_task = asyncio.create_task(
-            self._run_trigger_observation_queue_task(),
-            name="dynamic-scheduler-trigger-obs-queue",
-        )
-
-        self._cleanup_volume_removal_services_task = asyncio.create_task(
-            self._cleanup_volume_removal_services(),
-            name="dynamic-scheduler-cleanup-volume-removal-services",
-        )
-        await self._discover_running_services()
-
-    async def shutdown(self):
-        logger.info("Shutting down dynamic-sidecar scheduler")
-        self._keep_running = False
-        self._inverse_search_mapping = {}
-        self._to_observe = {}
-
-        if self._cleanup_volume_removal_services_task is not None:
-            self._cleanup_volume_removal_services_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await self._cleanup_volume_removal_services_task
-            self._cleanup_volume_removal_services_task = None
-
-        if self._scheduler_task is not None:
-            self._scheduler_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await self._scheduler_task
-            self._scheduler_task = None
-
-        if self._trigger_observation_queue_task is not None:
-            await self._trigger_observation_queue.put(None)
-
-            self._trigger_observation_queue_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await self._trigger_observation_queue_task
-            self._trigger_observation_queue_task = None
-            self._trigger_observation_queue = Queue()
-
-        # let's properly cleanup remaining observation tasks
-        running_tasks = self._service_observation_task.values()
-        for task in running_tasks:
-            task.cancel()
-        try:
-            MAX_WAIT_TIME_SECONDS = 5
-            results = await asyncio.wait_for(
-                asyncio.gather(*running_tasks, return_exceptions=True),
-                timeout=MAX_WAIT_TIME_SECONDS,
-            )
-            if bad_results := list(filter(lambda r: isinstance(r, Exception), results)):
-                logger.error(
-                    "Following observation tasks completed with an unexpected error:%s",
-                    f"{bad_results}",
-                )
-        except asyncio.TimeoutError:
-            logger.error(
-                "Timed-out waiting for %s to complete. Action: Check why this is blocking",
-                f"{running_tasks=}",
-            )
